requirements.txt
cryptography==39.0.2 
Pillow==10.0.0 
python-cfonts==1.4.0 
